Parse scores

+++ webapp/db/migrate/20140224101230_add_total_to_team.rb
@@ -1,5 +1,5 @@
 class AddTotalToTeam < ActiveRecord::Migration
   def change
-    add_column :teams, :total, :integer, index: true
+    add_column :teams, :total, :decimal, index: true
   end
 end
